requirements.txt
Django==5.0
django-simple-history==3.7.0